+++ CHANGELOG.md
@@ -61,7 +61,7 @@ To learn more about active deprecations, we recommend checking [GitHub Discussio
 
 ### New
 
-- TODO ([#XXX](https://github.com/kedacore/keda/issues/XXX))
+- **General**: Introduce new Weather-Aware Demand Scaler ([#6799](https://github.com/kedacore/keda/pull/6799))
 
 #### Experimental